Address PR #19 review comments (GDACS/ADAM)
Fixes from Tristan's review:

- storm_id_lookup: column-scoped upsert (src/pipelines/_upsert.py) so ADAM
  and match writes no longer NULL each other's columns — the all-columns
  stratus.postgres_upsert clobber. Wired into match/adam/gdacs; no-DB
  regression test; ADR 0001 Confirmation corrected. [#1,#2]
- validate_crosswalk.py: generalized to both sources, validates the
  humanreview CSVs (the real builder input), real exit-code gate; both
  builders now gate --write-db on it. [#4]
- adam: per-event exposure upsert for crash-durable backfills. [#3]
- FM lookups: BIGINT key + 4-col UNIQUE NULLS NOT DISTINCT in the DDL;
  builders TRUNCATE+append into the pre-created typed table (no more
  if_exists="replace"); dedicated init_db_{adam,gdacs}_lookup.py; gdacs
  join-doc corrected (gdacs_admin_code = gmi_admin). [#7,#8,#9]
- Archived v1 builders to scripts/_archive/; renamed *_v2 builders to the
  canonical names; repointed all references. [#6]
- Matchers marked EXPLORATORY/REVIEW-ONLY + README code map; IoU tie-break
  left as-is (review-only, not in the production path). [#10]
- Documented the gdacs_fm_lookup double-count contract for consumers. [#11]

ocha-lens pin alignment [#5] deferred to a tagged ocha-lens release.
