+++ src/schemas/sql/adam_fm_lookup.sql
@@ -5,26 +5,23 @@
 -- FieldMaps (FM) p-code, so ADAM exposure can be reported against
 -- standardized FM admin boundaries.
 --
--- Built offline, NOT by the runtime pipelines: scripts/build_adam_fm_lookup_v2.py
--- runs the spatial matcher in src/static/adam/matcher.py. It uses the
--- PER-ADAM-admin iteration (match_per_adam_admin): one row per ADAM admin
--- polygon, each picking its best-IoU FM polygon. That direction is deliberate
--- — ADAM exposure joins by admin NAME, and per-ADAM iteration guarantees each
--- adam_admin_id appears at most once, so the downstream name join stays 1:1
--- and can't fan out / double-count.
+-- Built offline, NOT by the runtime pipelines: scripts/build_adam_fm_lookup.py
+-- reads the human-reviewed crosswalk
+-- (data/review/adam_fm_crosswalk_humanreview.csv — the source of truth for
+-- reviewer decisions) and projects it into these lookup rows. The crosswalk
+-- itself is generated by scripts/build_adam_fm_crosswalk.py. (The spatial
+-- matcher in src/static/adam/matcher.py is REVIEW-ONLY and is NOT used by this
+-- build.) Rebuild via scripts/init_db_adam_lookup.py (drop+recreate) then the
+-- builder; rebuild only when FieldMaps or ADAM change their admin schema.
 --
 -- Downstream join: storms.adam_exposure.admin_name = adam_fm_lookup.adam_admin_name
 -- to attach (fm_pcode, fm_name) → adam_exposure.pcode.
 --
--- ----------------------------------------------------------------------------
--- NOTE on shape:
---
--- Created by a pandas to_sql write, so no constraints/indexes and every
--- column is nullable. The CREATE below faithfully transcribes that live
--- shape. A recommended natural key + index are included COMMENTED OUT.
+-- Natural key: (iso3, admin_level, fm_pcode, adam_admin_id), NULLS NOT
+-- DISTINCT so an FM-only row (NULL adam_admin_id) can't silently duplicate.
 -- ============================================================================
 
--- DROP TABLE IF EXISTS storms.adam_fm_lookup CASCADE;
+DROP TABLE IF EXISTS storms.adam_fm_lookup CASCADE;
 
 CREATE TABLE IF NOT EXISTS storms.adam_fm_lookup
 (
@@ -32,11 +29,13 @@ CREATE TABLE IF NOT EXISTS storms.adam_fm_lookup
     admin_level     BIGINT,
     fm_pcode        TEXT,
     fm_name         TEXT,
-    adam_admin_id   DOUBLE PRECISION,
+    adam_admin_id   BIGINT,
     adam_admin_name TEXT,
     iou             DOUBLE PRECISION,
     caveat_kind     TEXT,
-    caveat_note     TEXT
+    caveat_note     TEXT,
+    CONSTRAINT adam_fm_lookup_unique
+        UNIQUE NULLS NOT DISTINCT (iso3, admin_level, fm_pcode, adam_admin_id)
 )
 TABLESPACE pg_default;
 
@@ -44,7 +43,7 @@ ALTER TABLE IF EXISTS storms.adam_fm_lookup
     OWNER to {owner};
 
 COMMENT ON TABLE storms.adam_fm_lookup IS
-    'Static ADAM-admin → FieldMaps-pcode crosswalk. Built offline by scripts/build_adam_fm_lookup_v2.py via per-ADAM-admin IoU matching; join to storms.adam_exposure on admin_name.';
+    'Static ADAM-admin → FieldMaps-pcode crosswalk. Built offline by scripts/build_adam_fm_lookup.py from the human-reviewed crosswalk; join to storms.adam_exposure on admin_name.';
 COMMENT ON COLUMN storms.adam_fm_lookup.iso3 IS
     'ISO3 country code the admin unit belongs to.';
 COMMENT ON COLUMN storms.adam_fm_lookup.admin_level IS
@@ -54,7 +53,7 @@ COMMENT ON COLUMN storms.adam_fm_lookup.fm_pcode IS
 COMMENT ON COLUMN storms.adam_fm_lookup.fm_name IS
     'FieldMaps admin name for fm_pcode.';
 COMMENT ON COLUMN storms.adam_fm_lookup.adam_admin_id IS
-    'ADAM-native admin id (stored as double precision as written). Unique per row by construction (per-ADAM iteration).';
+    'ADAM-native admin id (BIGINT). Part of the natural key; NULL on FM-only rows.';
 COMMENT ON COLUMN storms.adam_fm_lookup.adam_admin_name IS
     'ADAM admin name — the join key against storms.adam_exposure.admin_name.';
 COMMENT ON COLUMN storms.adam_fm_lookup.iou IS
@@ -64,10 +63,5 @@ COMMENT ON COLUMN storms.adam_fm_lookup.caveat_kind IS
 COMMENT ON COLUMN storms.adam_fm_lookup.caveat_note IS
     'Free-text detail accompanying caveat_kind.';
 
--- ============================================================================
--- Recommended (not yet applied in the live DB) — uncomment in a migration
--- ============================================================================
--- ALTER TABLE storms.adam_fm_lookup
---     ADD CONSTRAINT adam_fm_lookup_unique UNIQUE (iso3, admin_level, adam_admin_id);
--- CREATE INDEX IF NOT EXISTS idx_adam_fm_lookup_admin_name
---     ON storms.adam_fm_lookup (adam_admin_name) WHERE adam_admin_name IS NOT NULL;
+CREATE INDEX IF NOT EXISTS idx_adam_fm_lookup_admin_name
+    ON storms.adam_fm_lookup (adam_admin_name) WHERE adam_admin_name IS NOT NULL;
