+++ src/schemas/sql/gdacs_fm_lookup.sql
@@ -5,25 +5,33 @@
 -- FieldMaps (FM) p-code, so GDACS exposure (keyed on GDACS-native admin
 -- codes) can be reported against standardized FM admin boundaries.
 --
--- Built offline, NOT by the runtime pipelines: scripts/build_gdacs_fm_lookup_v2.py
--- runs the spatial matcher in src/static/gdacs/matcher.py (per-FM-unit IoU
--- against GDACS admin polygons) and writes the result here. It is rebuilt
--- only when FieldMaps or GDACS change their admin schema.
+-- Built offline, NOT by the runtime pipelines: scripts/build_gdacs_fm_lookup.py
+-- reads the human-reviewed crosswalk
+-- (data/review/gdacs_fm_crosswalk_humanreview.csv — the source of truth for
+-- reviewer decisions) and projects it into these lookup rows. The crosswalk
+-- itself is generated by scripts/build_gdacs_fm_crosswalk.py. (The spatial
+-- matcher in src/static/gdacs/matcher.py is REVIEW-ONLY and is NOT used by
+-- this build.) Rebuild via scripts/init_db_gdacs_lookup.py (drop+recreate)
+-- then the builder; rebuild only when FieldMaps or GDACS change their schema.
 --
--- Downstream join: storms.gdacs_exposure.gmi_admin = gdacs_fm_lookup.gmi_admin
--- to attach (fm_pcode, fm_name).
+-- Downstream join: storms.gdacs_exposure.gdacs_admin_code = gdacs_fm_lookup.gmi_admin
+-- to attach (fm_pcode, fm_name). NOTE the column names differ across the two
+-- sides (gdacs_exposure stores the native code as `gdacs_admin_code`, this
+-- table as `gmi_admin`); the consumer maps them explicitly. A symmetric
+-- rename is deferred (cross-repo: ds-storms-alerts joins on these names).
 --
--- ----------------------------------------------------------------------------
--- NOTE on shape:
+-- DOUBLE-COUNT CONTRACT: a single gmi_admin (coarse GDACS unit) may map to
+-- MULTIPLE fm_pcode rows here (many-FM ← one-GDACS). The exposure→lookup join
+-- therefore fans out, so any consumer that sums population MUST aggregate by
+-- fm_pcode (e.g. groupby('fm_pcode').sum('pop_exposed')) or it will
+-- double-count. Consumer lives in ds-storms-alerts (exposure_enrich.py) —
+-- verify it honors this (tracked as a follow-up; see PR #19).
 --
--- The table is currently created by a pandas to_sql write, so it carries no
--- constraints or indexes and every column is nullable. The CREATE below is a
--- faithful transcription of that live shape. A recommended natural key and
--- index are included COMMENTED OUT — apply them in a deliberate migration if
--- this table starts being joined at scale.
+-- Natural key: (iso3, admin_level, fm_pcode, gmi_admin), NULLS NOT DISTINCT
+-- so an FM-only row (NULL gmi_admin) can't silently duplicate.
 -- ============================================================================
 
--- DROP TABLE IF EXISTS storms.gdacs_fm_lookup CASCADE;
+DROP TABLE IF EXISTS storms.gdacs_fm_lookup CASCADE;
 
 CREATE TABLE IF NOT EXISTS storms.gdacs_fm_lookup
 (
@@ -34,7 +42,9 @@ CREATE TABLE IF NOT EXISTS storms.gdacs_fm_lookup
     gmi_admin        TEXT,
     gdacs_admin_name TEXT,
     caveat_kind      TEXT,
-    caveat_note      TEXT
+    caveat_note      TEXT,
+    CONSTRAINT gdacs_fm_lookup_unique
+        UNIQUE NULLS NOT DISTINCT (iso3, admin_level, fm_pcode, gmi_admin)
 )
 TABLESPACE pg_default;
 
@@ -42,7 +52,7 @@ ALTER TABLE IF EXISTS storms.gdacs_fm_lookup
     OWNER to {owner};
 
 COMMENT ON TABLE storms.gdacs_fm_lookup IS
-    'Static GDACS-admin → FieldMaps-pcode crosswalk. Built offline by scripts/build_gdacs_fm_lookup_v2.py via spatial (IoU) matching; join to storms.gdacs_exposure on gmi_admin.';
+    'Static GDACS-admin → FieldMaps-pcode crosswalk. Built offline by scripts/build_gdacs_fm_lookup.py from the human-reviewed crosswalk; join storms.gdacs_exposure.gdacs_admin_code to gdacs_fm_lookup.gmi_admin.';
 COMMENT ON COLUMN storms.gdacs_fm_lookup.iso3 IS
     'ISO3 country code the admin unit belongs to.';
 COMMENT ON COLUMN storms.gdacs_fm_lookup.admin_level IS
@@ -52,7 +62,7 @@ COMMENT ON COLUMN storms.gdacs_fm_lookup.fm_pcode IS
 COMMENT ON COLUMN storms.gdacs_fm_lookup.fm_name IS
     'FieldMaps admin name for fm_pcode.';
 COMMENT ON COLUMN storms.gdacs_fm_lookup.gmi_admin IS
-    'GDACS-native GMI admin code — the join key against storms.gdacs_exposure.';
+    'GDACS-native GMI admin code — the join key, matched against storms.gdacs_exposure.gdacs_admin_code (note the differing column names). Part of the natural key; NULL on FM-only rows.';
 COMMENT ON COLUMN storms.gdacs_fm_lookup.gdacs_admin_name IS
     'GDACS admin name for gmi_admin.';
 COMMENT ON COLUMN storms.gdacs_fm_lookup.caveat_kind IS
@@ -60,10 +70,5 @@ COMMENT ON COLUMN storms.gdacs_fm_lookup.caveat_kind IS
 COMMENT ON COLUMN storms.gdacs_fm_lookup.caveat_note IS
     'Free-text detail accompanying caveat_kind (e.g. which units were aggregated).';
 
--- ============================================================================
--- Recommended (not yet applied in the live DB) — uncomment in a migration
--- ============================================================================
--- ALTER TABLE storms.gdacs_fm_lookup
---     ADD CONSTRAINT gdacs_fm_lookup_unique UNIQUE (iso3, admin_level, gmi_admin);
--- CREATE INDEX IF NOT EXISTS idx_gdacs_fm_lookup_gmi_admin
---     ON storms.gdacs_fm_lookup (gmi_admin) WHERE gmi_admin IS NOT NULL;
+CREATE INDEX IF NOT EXISTS idx_gdacs_fm_lookup_gmi_admin
+    ON storms.gdacs_fm_lookup (gmi_admin) WHERE gmi_admin IS NOT NULL;
