+++ src/static/adam/matcher.py
@@ -1,15 +1,37 @@
 """Spatial matching: FieldMaps adm-1 polygons ↔ ADAM admin polygons.
 
+============================================================================
+⚠️  EXPLORATORY / REVIEW-ONLY — NOT live production monitoring.
+----------------------------------------------------------------------------
+This module is NOT used by the production lookup or the real-time pipelines.
+The production `storms.adam_fm_lookup` is built by
+`scripts/build_adam_fm_lookup.py` from the human-reviewed crosswalk
+(`data/review/adam_fm_crosswalk_humanreview.csv`); that crosswalk is
+generated by `scripts/build_adam_fm_crosswalk.py`, which runs its own
+overlay and emits EVERY FM×ADAM pair — it does not call this module.
+
+Only caller after this PR: the human-review report pipeline
+(`scripts/build_adam_review_artifacts.py` → `exploration/review_report_adam
+.qmd`), which uses `match_country`. (The legacy v1 builder — now archived at
+`scripts/_archive/build_adam_fm_lookup.py` — used both functions; with it
+archived, `match_per_adam_admin` has no caller.)
+
+Consequence: the IoU best-pick tie-break below is review-only and left as-is
+on purpose (see PR #19 discussion). A later PR will relocate this out of
+`src/static/` so the exploratory↔production boundary is structural.
+============================================================================
+
 Mirror of :mod:`src.static.gdacs.matcher` for the FM↔ADAM bridge.
 Two iteration modes:
 
   - :func:`match_country` — iterates the FM side, picks best-IoU ADAM
-    polygon per FM unit. Used historically; kept for the review
-    artifacts builder (per-FM diagnostics).
+    polygon per FM unit. Used by the review-artifacts builder (per-FM
+    diagnostics).
   - :func:`match_per_adam_admin` — iterates the ADAM side, picks
-    best-IoU FM polygon per ADAM polygon. Used by the lookup build to
-    eliminate the name-fanout class of bug (each ADAM admin_id appears
-    at most once in the lookup, so name joins downstream resolve 1:1).
+    best-IoU FM polygon per ADAM polygon. Was used by the v1 lookup build
+    (archived in PR #19) to avoid name-fanout double-counting; currently
+    has no caller (kept for the per-source approach a future production
+    rebuild may adopt).
 
 The ADAM admin layer is fixed at admin-1 (no hierarchy); only the
 FM-side level is configurable via
