+++ src/static/gdacs/matcher.py
@@ -1,5 +1,25 @@
 """Spatial matching: FieldMaps adm-1 polygons ↔ GDACS admin polygons.
 
+============================================================================
+⚠️  EXPLORATORY / REVIEW-ONLY — NOT live production monitoring.
+----------------------------------------------------------------------------
+This module is NOT used by the production lookup or the real-time pipelines.
+The production `storms.gdacs_fm_lookup` is built by
+`scripts/build_gdacs_fm_lookup.py` from the human-reviewed crosswalk
+(`data/review/gdacs_fm_crosswalk_humanreview.csv`); that crosswalk is
+generated by `scripts/build_gdacs_fm_crosswalk.py`, which runs its own
+overlay and emits EVERY FM×GDACS pair — it does not call this module.
+
+Only caller after this PR: the human-review report pipeline
+(`scripts/build_review_artifacts.py` → `exploration/review_report_gdacs.qmd`).
+(The legacy v1 builder — now archived at
+`scripts/_archive/build_canonical_lookup.py` — also used it.)
+
+Consequence: the IoU best-pick tie-break below is review-only and left as-is
+on purpose (see PR #19 discussion). A later PR will relocate this out of
+`src/static/` so the exploratory↔production boundary is structural.
+============================================================================
+
 Pairs each FieldMaps unit with the GDACS unit that maximises
 intersection-over-union (IoU) inside the country. Outputs one row per
 FieldMaps polygon — even when no GDACS polygon overlaps it (those rows
@@ -9,9 +29,10 @@ GDACS's admin layer is fixed at admin-1 (no hierarchy) so only the
 FM-side level is configurable per country (via
 :func:`src.static.gdacs.inputs.resolve_gdacs_fm_level`).
 
-This is the one expensive step in the lookup build. Called once per
-country by ``scripts/build_canonical_lookup.py``; the whole basin
-takes ~minutes on a laptop.
+This is the expensive step. Called once per country by
+``scripts/build_review_artifacts.py`` (the review-report pipeline; see the
+banner above) — and historically by the archived v1 lookup builder; the
+whole basin takes ~minutes on a laptop.
 """
 
 from __future__ import annotations
