+++ tests/conftest.py
@@ -1,0 +1,6 @@
+import sys
+from pathlib import Path
+
+# Make the repo root importable so tests can `import src...` the same way
+# run_pipeline.py does.
+sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
